Fixed several issues with the async query provider and missing results due to the stored queries not being arranged properly

+++ SanteDB.Core.Model/EntityLoader/EntitySource.cs
@@ -54,7 +54,7 @@ namespace SanteDB.Core.Model.EntityLoader
             /// <summary>
             /// Gets the specified relations
             /// </summary>
-            public IEnumerable<TObject> GetRelations<TObject>(Guid? sourceKey, decimal? sourceVersionSequence) where TObject : IdentifiedData, IVersionedAssociation, new()
+            public IEnumerable<TObject> GetRelations<TObject>(Guid? sourceKey, int? sourceVersionSequence) where TObject : IdentifiedData, IVersionedAssociation, new()
             {
                 return new List<TObject>();
             }
